refactor(m9): extraire SummonAnimations — SummonParticles, SummonExplosion, HeroRevealCard (#267)
- Crée src/components/summon/SummonAnimations.tsx qui exporte rarityGlows,
  SummonParticles, SummonExplosion et HeroRevealCard
- HeroRevealCard unifié avec prop variant ('modal'|'page') pour gérer
  HeroAvatar (SummonModal) vs PixelIcon (page Summon)
- SummonModal.tsx et Summon.tsx importent depuis le nouveau module
- Supprime les imports devenus inutiles (useEffect, ChevronLeft/Right dans
  SummonModal ; rarityGlows dans Summon)
